[refactor] #53 - 이미지 업로드 컴포넌트 분리
